+++ lib/neuron.ex
@@ -87,8 +87,9 @@ defmodule Neuron do
   ```
   """
 
-  @spec query(query_string :: String.t(), variables :: Map.t(), options :: keyword()) ::
-          {:ok, Neuron.Response.t()} | {:error, Neuron.Response.t() | Neuron.JSONParseError.t()}
+  @spec query(query_string :: String.t(), variables :: map(), options :: keyword()) ::
+          {:ok, Neuron.Response.t()}
+          | {:error, Neuron.Response.t() | Neuron.JSONParseError.t() | HTTPoison.Error.t()}
   def query(query_string, variables \\ %{}, options \\ []) do
     json_library = json_library(options)